requirements.txt
absl-py==1.2.0
anyio==2.2.0
appdirs==1.4.4
argon2-cffi==20.1.0
args==0.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
audioread==2.1.9
auraloss==0.2.1
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
cached-property==1.5.2
calmsize==0.1.3
certifi==2020.6.20
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.12
click==7.1.2
clint==0.5.1
contextvars==2.4
cupy-cuda102==8.6.0
cvxopt==1.3.0
cycler==0.11.0
Cython==0.29.32
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
docopt==0.6.2
entrypoints==0.3
enum34==1.1.10
essentia==2.1b6.dev374
fastrlock==0.6
findiff==0.8.9
future==0.18.2
gast==0.5.3
graphviz==0.17
grpcio==1.47.0
h5py==3.1.0
idna==3.1
imageio==2.9.0
immutables==0.15
importlib-metadata==4.8.3
ipykernel==5.5.3
ipython==7.16.1
itsdangerous==2.0.1
jams==0.3.4
jedi==0.18.0
Jinja2==3.0.3
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.9.1
jupyter-server==1.6.1
jupyterlab==3.0.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
librosa==0.9.2
llvmlite==0.36.0
madmom==0.16.1
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.2
mido==1.2.10
mir-eval==0.7
mistune==0.8.4
mock==4.0.3
mpmath==1.2.1
msaf==0.1.70
mutagen==1.45.1
nbclassic==0.2.7
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.3.0
numba==0.53.1
numpy==1.18.1
nvidia-ml-py==11.450.51
openunmix==1.2.1
packaging==21.3
pandas==1.1.3
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
pipenv==2020.11.15
pooch==1.6.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.19.4
ptyprocess==0.7.0
PyAudio==0.2.11
pycparser==2.21
Pygments==2.8.1
pyloudnorm==0.1.0
pyparsing==3.0.9
pyrsistent==0.18.0
pyrubberband==0.3.0
PySoundFile==0.9.0.post1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pyzmq==22.0.3
requests==2.27.1
resampy==0.4.0
scdl==1.6.12
scikit-image==0.17.2
scikit-learn==0.21.3
scipy==1.2.1
seaborn==0.11.2
Send2Trash==1.5.0
sh==1.14.1
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
sympy==1.8
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
tifffile==2020.9.3
tomlkit==0.7.0
torch==1.10.1
torchaudio==0.10.1
torchvision==0.5.0
tornado==6.1
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.11
virtualenv-clone==0.5.4
vmo==0.23.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
youtube-dl==2021.6.6
zipp==3.4.1
